+++ config/enums-plus.php
@@ -1,6 +1,20 @@
 <?php
 
-// config for Rea/LaravelEnumsPlus
+// Config for Rea/LaravelEnumsPlus
 return [
+
+    /**
+     * Define the default path to load translations from.
+     */
+    'translations' => 'enums',
+
+    /**
+     * Define the column names used by the selection() method.
+     */
+    'columns' => [
+        'value' => 'value',
+        'label' => 'label',
+        'selected' => 'selected',
+    ],
 
 ];